TurkceRumenceCeviri/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-language-spellcheck==4.0.0
pytesseract==0.3.10
pillow==10.0.0
opencv-python==4.8.0.74
requests==2.31.0
transformers==4.33.0
torch==2.0.0
